refactor: extract EmailVerificationViewModel and add verification protocols
Mirrors the phone extraction: EmailVerificationViewModel owns email
state, actions, deeplink handling, and email-specific dialogs.
OnrampVerificationViewModel becomes a thin orchestrator composing both
verifiers via callbacks (onCodeRequested → push confirm screen,
onVerified → advance to next step). PhoneVerifying and EmailVerifying
protocols define the host-facing contract that both VMs conform to,
unlocking substitution at the construction site.

Renames resetTransientState → reset on the protocol surface. Moves
email screens to Screens/EmailVerification/ to match the phone
folder layout.
