+++ FlipcashTests/OnrampVerificationViewModelTests.swift
@@ -11,11 +11,11 @@ import FlipcashCore
 @Suite("OnrampVerificationViewModel")
 struct OnrampVerificationViewModelTests {
 
-    @Test("applyDeeplinkVerification sets the confirm-email path and stashes the email")
-    func applyDeeplinkVerification_setsConfirmEmailPath() {
+    @Test("applyDeeplinkVerification forwards to the inner email VM and resets Onramp's path")
+    func applyDeeplinkVerification_forwardsAndSetsPath() {
         let viewModel = OnrampVerificationViewModel(session: .unverifiedMock, flipClient: .mock)
         viewModel.applyDeeplinkVerification(VerificationDescription(email: "a@b.c", code: "123456"))
         #expect(viewModel.verificationPath == [.confirmEmailCode])
-        #expect(viewModel.enteredEmail == "a@b.c")
+        #expect(viewModel.emailViewModel.enteredEmail == "a@b.c")
     }
 }
